requirements.txt
pandas
plotly
streamlit
kafka-python
